# public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BetAki Admin API",
        "description": "API REST da Dashboard de Administração da Betaki (Laravel 12 / PHP 8.4).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Ambiente atual"
        }
    ],
    "paths": {
...
        "/api/v1/footers/{footer}/publish": {
            "post": {
                "tags": [
                    "Footers"
                ],
                "summary": "Publicar footer",
                "description": "Publicar footer",
                "operationId": "8cfdf176194df5e3e79e85826c7fac81",
                "parameters": [
                    {
                        "name": "footer",
                        "in": "path",
                        "description": "ID do footer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Footer publicado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "key": {
                                            "type": "string",
                                            "example": "main-footer"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "published"
                                        },
                                        "publish_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "published_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Footer não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/footers/{footer}/links": {
            "put": {
                "tags": [
                    "Footers"
                ],
                "summary": "Sincronizar links de um footer",
                "description": "Sincronizar links do footer",
                "operationId": "3347730f376fa8c75d898beeb081f0ec",
                "parameters": [
                    {
                        "name": "footer",
                        "in": "path",
                        "description": "ID do footer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "links"
                                ],
                                "properties": {
                                    "links": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5,
                                                    "nullable": true
                                                },
                                                "block": {
                                                    "type": "string",
                                                    "example": "legal",
                                                    "nullable": true
                                                },
                                                "label": {
                                                    "type": "string",
                                                    "example": "Política de Privacidade"
                                                },
                                                "url": {
                                                    "type": "string",
                                                    "example": "/privacy-policy"
                                                },
                                                "icon": {
                                                    "type": "string",
                                                    "example": "instagram",
                                                    "nullable": true
                                                },
                                                "target": {
                                                    "type": "string",
                                                    "example": "_self"
                                                },
                                                "position": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Links sincronizados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "key": {
                                            "type": "string",
                                            "example": "main-footer"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "draft"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Footer não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Envie: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}